@@ -17,11 +17,10 @@
 package com.stasbar.prompter
 
 import android.content.Context
-import android.support.v4.app.FragmentManager
-import android.support.v7.app.AppCompatActivity
 import android.text.InputType
 import android.view.View
 import android.widget.TextView
+import androidx.fragment.app.FragmentManager
 
 /**
  * Created by stasbar on 31.10.2017
